Rename TestMutablePreferenceImpl as MutablePreferenceImplTest and add GivenTest
